pyproject.toml
[project]
name = "aarambh"
version = "0.1.0"
description = "Emergency Services Voice Agent & RAG Chatbot"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # --- Core Application ---
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    # --- Voice Processing ---
    "websockets>=12.0",
    "twilio>=9.0.0",
    "deepgram-sdk>=3.0.0",
    "ffmpeg-python>=0.2.0",
    "requests>=2.31.0",
    # --- LLM & AI ---
    "google-genai>=0.2.0", # Google's official GenAI SDK (for Gemini 1.5/2.0)
    "langchain-google-genai>=2.0.0", # LangChain integration for Google
    # --- LangChain Ecosystem (0.3.x) ---
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langchainhub>=0.1.20",
    # --- RAG & Vector Stores ---
    "langchain-chroma>=0.1.2",
    "chromadb>=0.5.0",
    "langchain-pinecone>=0.2.0",
    "pinecone-client>=5.0.0",
    "langchain-huggingface>=0.1.0",
    # --- Retrieval Optimization ---
    "sentence-transformers>=3.0.0",
    "flashrank>=0.2.0",
    "rank-bm25>=0.2.2",
    # --- Utilities ---
    "pymupdf>=1.24.10",
    "geopy>=2.4.1",
    "streamlit>=1.51.0",
]

[tool.uv]
compile-bytecode = true
